homework2/kafka_cons_bl/requirements.txt
attrs==20.3.0
importlib-metadata==3.0.0
jsonschema==3.2.0
kafka-python==2.0.2
numpy==1.19.4
pandas==1.1.4
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
zipp==3.4.0
